+++ package.json
@@ -47,5 +47,8 @@
     "react-dom": "^18.3.1",
     "tailwindcss": "^3.4.4",
     "vite": "^5.3.1"
+  },
+  "build": {
+    "productName": "Stellar"
   }
 }
