--- src/renderer/src/home/Dashboard.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-function Dashboard() {
-  return <div>Dashboard</div>
-}
-
-export default Dashboard
