+++ src/renderer/src/home/Reports.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function Reports() {
+  return (
+    <div>Reports</div>
+  )
+}
+
+export default Reports
