+++ src/renderer/src/home/Tax.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function Tax() {
+  return (
+    <div>Tax</div>
+  )
+}
+
+export default Tax
